core/dna: extract Chromosome/DNA validation into @holons/core
Move Holon DNA domain (types, validation, holosphere persistence,
seed data) out of apps/web/src/lib/dna into packages/core/src/dna so
that future telegram, AI, and text UIs can operate on the same
contracts without duplication. Web sources become thin re-exports
from @holons/core/dna; runtime behavior is preserved.

While porting persistence, fold the read-then-merge double fetch in
addChromosome/updateChromosome into a single library read.
