+++ packages/core/package.json
@@ -17,6 +17,7 @@
     "holosphere": "1.3.0-alpha4"
   },
   "devDependencies": {
+    "@types/node": "^22.15.17",
     "typescript": "^5.7.3",
     "vitest": "^4.0.18"
   }
